Removed the bug in main.tf of Cloudwatch_Logs_To_S3.
It turned out that the policy being created has the same name as the
policy that was created by Cloudtrail_logs_To_S3. There is still a bug
in the lambda function inside lambda_payload.zip which is due to the
limit in AWS for having only one export task at a time.